Refactoring and implementation of baseline
This commit refactors the existing code to utilize OOP features of C++:
The different calculation modes are represented through classes deriving
from a common base class `SummationStrategy`. This class handles the
distribution of summands and has a common interface to calculate the
sum.

Also, the so called `BaselineSummation` strategy has been implemented.
It consists of a simple MPI_Gather to the first rank, a local summation
using a std::accumulate followed by a MPI_Broadcast of the result.